+++ packages/angular/src/index.ts
@@ -23,6 +23,7 @@ export * from './tree-chart.component';
 export * from './treemap-chart.component';
 export * from './circle-pack-chart.component';
 export * from './wordcloud-chart.component';
+export * from './alluvial-chart.component';
 
 // Diagrams
 export * from './diagrams/card-node/card-node.module';