+++ packages/svelte/src/index.js
@@ -20,6 +20,7 @@ import TreeChart from './TreeChart.svelte';
 import TreemapChart from './TreemapChart.svelte';
 import CirclePackChart from './CirclePackChart.svelte';
 import WordCloudChart from './WordCloudChart.svelte';
+import AlluvialChart from './AlluvialChart.svelte';
 
 export {
 	AreaChart,
@@ -44,4 +45,5 @@ export {
 	TreemapChart,
 	CirclePackChart,
 	WordCloudChart,
+	AlluvialChart,
 };